Allow status info to be copied to the clipboard

+++ Almond/Sensor/SFISensorsViewController.m
@@ -805,7 +805,7 @@
     }
 
     if (newDeviceList.count != newDeviceValueList.count) {
-        ELog(@"Warning: device list and values lists are incongruent, d:%ld, v:%ld", (unsigned long)newDeviceList.count, newDeviceValueList.count);
+        ELog(@"Warning: device list and values lists are incongruent, d:%ld, v:%ld", (unsigned long)newDeviceList.count, (unsigned long)newDeviceValueList.count);
     }
 
     DLog(@"Changing device value list: %@", newDeviceValueList);
